+++ src/main/java/edu/harvard/iq/dataverse/MetadataBlockDao.java
@@ -1,7 +1,6 @@
 package edu.harvard.iq.dataverse;
 
 import javax.ejb.Stateless;
-import javax.inject.Named;
 import javax.persistence.EntityManager;
 import javax.persistence.NoResultException;
 import javax.persistence.PersistenceContext;
@@ -12,7 +11,6 @@ import java.util.List;
  * @author michael
  */
 @Stateless
-@Named
 public class MetadataBlockDao {
     
     @PersistenceContext(unitName = "VDCNet-ejbPU")
