+++ src/main/java/edu/harvard/iq/dataverse/api/DatasetFieldServiceApi.java
@@ -157,7 +157,7 @@ public class DatasetFieldServiceApi extends AbstractApiBean {
                     .add("parentAllowsMultiples", parentAllowsMultiplesDisplay)
                     .add("solrFieldSearchable", solrFieldSearchable)
                     .add("solrFieldFacetable", solrFieldFacetable)
-                    .add("isCitationMetaBlock", isRequired));
+                    .add("isRequired", isRequired));
         
         } catch ( NoResultException nre ) {
             return notFound(name);
