+++ src/main/java/edu/harvard/iq/dataverse/dataverse/DataverseSaver.java
@@ -1,0 +1,142 @@
+package edu.harvard.iq.dataverse.dataverse;
+
+import com.google.common.collect.Lists;
+import edu.harvard.iq.dataverse.DatasetFieldType;
+import edu.harvard.iq.dataverse.Dataverse;
+import edu.harvard.iq.dataverse.DataverseFieldTypeInputLevel;
+import edu.harvard.iq.dataverse.DataverseRequestServiceBean;
+import edu.harvard.iq.dataverse.DataverseServiceBean;
+import edu.harvard.iq.dataverse.DataverseSession;
+import edu.harvard.iq.dataverse.EjbDataverseEngine;
+import edu.harvard.iq.dataverse.UserNotification;
+import edu.harvard.iq.dataverse.UserNotificationServiceBean;
+import edu.harvard.iq.dataverse.authorization.users.AuthenticatedUser;
+import edu.harvard.iq.dataverse.authorization.users.User;
+import edu.harvard.iq.dataverse.engine.command.Command;
+import edu.harvard.iq.dataverse.engine.command.exception.CommandException;
+import edu.harvard.iq.dataverse.engine.command.impl.CreateDataverseCommand;
+import edu.harvard.iq.dataverse.engine.command.impl.UpdateDataverseCommand;
+import edu.harvard.iq.dataverse.error.DataverseError;
+import edu.harvard.iq.dataverse.settings.SettingsWrapper;
+import edu.harvard.iq.dataverse.util.BundleUtil;
+import edu.harvard.iq.dataverse.util.JsfHelper;
+import edu.harvard.iq.dataverse.util.SystemConfig;
+import io.vavr.control.Either;
+import org.primefaces.model.DualListModel;
+
+import javax.ejb.Stateless;
+import javax.faces.application.FacesMessage;
+import javax.inject.Inject;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.logging.Level;
+import java.util.logging.Logger;
+
+import static edu.harvard.iq.dataverse.util.JsfHelper.JH;
+
+@Stateless
+public class DataverseSaver {
+
+    private static final Logger logger = Logger.getLogger(DataverseSaver.class.getCanonicalName());
+
+    @Inject
+    private DataverseSession session;
+
+    @Inject
+    private DataverseRequestServiceBean dvRequestService;
+
+    @Inject
+    private DataverseServiceBean dataverseService;
+
+    @Inject
+    private EjbDataverseEngine commandEngine;
+
+    @Inject
+    private UserNotificationServiceBean userNotificationService;
+
+    @Inject
+    private SettingsWrapper settingsWrapper;
+
+    @Inject
+    private SystemConfig systemConfig;
+
+    // -------------------- LOGIC --------------------
+
+    public Either<DataverseError, Dataverse> saveNewDataverse(Collection<DataverseFieldTypeInputLevel> dftilToBeSaved,
+                                                              Dataverse dataverse,
+                                                              DualListModel<DatasetFieldType> facets) {
+
+        if (!dataverse.isFacetRoot()) {
+            facets.getTarget().clear();
+        }
+
+        if (session.getUser().isAuthenticated()) {
+            dataverse.setOwner(dataverse.getOwner().getId() != null ? dataverseService.find(dataverse.getOwner().getId()) : null);
+            Command<Dataverse> cmd = new CreateDataverseCommand(dataverse,
+                    dvRequestService.getDataverseRequest(),
+                    facets.getTarget(),
+                    Lists.newArrayList(dftilToBeSaved));
+
+            try {
+                dataverse = commandEngine.submit(cmd);
+            } catch (CommandException ex) {
+                logger.log(Level.SEVERE, "Unexpected Exception calling dataverse command", ex);
+                JH.addMessage(FacesMessage.SEVERITY_FATAL, BundleUtil.getStringFromBundle("dataverse.create.failure"));
+                return Either.left(new DataverseError(ex, BundleUtil.getStringFromBundle("dataverse.create.failure")));
+            }
+
+            sendSuccessNotificationAsync(dataverse, session.getUser());
+
+            showSuccessMessage();
+        } else {
+            JH.addMessage(FacesMessage.SEVERITY_FATAL, BundleUtil.getStringFromBundle("dataverse.create.authenticatedUsersOnly"));
+            return Either.left(new DataverseError(BundleUtil.getStringFromBundle("dataverse.create.authenticatedUsersOnly")));
+        }
+
+        return Either.right(dataverse);
+    }
+
+    public Either<DataverseError, Dataverse> saveEditedDataverse(Collection<DataverseFieldTypeInputLevel> dftilToBeSaved,
+                                                                 Dataverse dataverse,
+                                                                 DualListModel<DatasetFieldType> facets) {
+
+        if (!dataverse.isFacetRoot()) {
+            facets.getTarget().clear();
+        }
+
+        UpdateDataverseCommand cmd = new UpdateDataverseCommand(dataverse, facets.getTarget(), null,
+                dvRequestService.getDataverseRequest(), Lists.newArrayList(dftilToBeSaved));
+
+        try {
+            dataverse = commandEngine.submit(cmd);
+
+            JsfHelper.addFlashSuccessMessage(BundleUtil.getStringFromBundle("dataverse.update.success"));
+        } catch (CommandException ex) {
+            logger.log(Level.SEVERE, "Unexpected Exception calling dataverse command", ex);
+            String errMsg = BundleUtil.getStringFromBundle("dataverse.update.failure");
+            JH.addMessage(FacesMessage.SEVERITY_FATAL, errMsg);
+            return Either.left(new DataverseError(ex, BundleUtil.getStringFromBundle("dataverse.create.failure")));
+        }
+
+        return Either.right(dataverse);
+    }
+
+    // -------------------- PRIVATE --------------------
+
+    private void showSuccessMessage() {
+        JsfHelper.addFlashSuccessMessage(BundleUtil.getStringFromBundle("dataverse.create.success",
+                Arrays.asList(settingsWrapper.getGuidesBaseUrl(), systemConfig.getGuidesVersion())));
+    }
+
+    private void sendSuccessNotificationAsync(Dataverse dataverse, User user) {
+        ExecutorService executorService = Executors.newSingleThreadExecutor();
+
+        executorService.execute(() ->
+                userNotificationService.sendNotification((AuthenticatedUser) user, dataverse.getCreateDate(),
+                        UserNotification.Type.CREATEDV, dataverse.getId()));
+
+        executorService.shutdown();
+    }
+}
